dom/media/test/test_invalid_seek.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Media test: invalide seek test</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<video id='v'></video>
<pre id="test">
<script class="testbody" type="text/javascript">
// http://www.whatwg.org/specs/web-apps/current-work/#dom-media-seek
// If the media element's readyState is HAVE_NOTHING, then the user agent
// must raise an InvalidStateError exception.
var v = document.getElementById('v');
var passed = false;

ok(v.readyState == HTMLMediaElement.HAVE_NOTHING,
   "Invalid ready state in media element");

try {
  v.seek(1);
}
catch(e) {
  passed = true;
}

ok(passed, "Video did not raise error during invalid seek");
</script>
</pre>
</body>
</html>
